Separate progress bar function into its own file
